refactor(project): Refactor the project structure to account for testing
