+++ libgdx/android/src/com/darrandyford/AndroidLauncher.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
-import com.darrandyford.DarrAndyFord;
+import main.java.com.darrandyford.DarrAndyFord;
 
 public class AndroidLauncher extends AndroidApplication {
 	@Override
