--- libgdx/core/src/com/darrandyford/world/physics/PhysicsEntity.java
@@ -1,6 +1,0 @@
-package com.darrandyford.world.physics;
-
-public interface PhysicsEntity {
-	void initPhysics();
-	void updatePhysics(float deltaTime);
-}
