+++ libgdx/core/src/main/java/com/darrandyford/world/physics/PhysicsEntity.java
@@ -1,0 +1,6 @@
+package main.java.com.darrandyford.world.physics;
+
+public interface PhysicsEntity {
+	void initPhysics();
+	void updatePhysics(float deltaTime);
+}
