+++ libgdx/ios/src/com/darrandyford/IOSLauncher.java
@@ -5,7 +5,7 @@ import org.robovm.apple.uikit.UIApplication;
 
 import com.badlogic.gdx.backends.iosrobovm.IOSApplication;
 import com.badlogic.gdx.backends.iosrobovm.IOSApplicationConfiguration;
-import com.darrandyford.DarrAndyFord;
+import main.java.com.darrandyford.DarrAndyFord;
 
 public class IOSLauncher extends IOSApplication.Delegate {
     @Override
